Move `course.py` from `udemypy/udemy` to `udemypy`
Since `course.py` contains the data structure used in all udemypy
packages, it's more appropriate to place it at the `udemypy/`
level.
